api/requirements.txt
django
djangorestframework
django-filter
scikit-learn
django-cors-headers